az_ai_search_index_enrich_func_app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
 
azure-functions
openai
requests
pandas
numpy
flask
flask_restx
azure-search-documents
azure-identity
azure-storage-blob
python-dotenv
pytest
azure-keyvault-secrets